+++ NEWS.md
@@ -80,6 +80,8 @@
   variables between the data frames, and also give a better error message when
   only one data frame has a by variable (#371). 
 
+* `top_n` correctly returns `n` rows instead of `n - 1` (#367).
+
 # dplyr 0.1.3
 
 ## Bug fixes